Move Header and Footer to templates
