+++ src/templates/Footer/index.jsx
@@ -2,21 +2,20 @@
 
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import Wrapper from './Wrapper';
 import Menu from '../../components/Menu';
-import Branding from './Branding';
 
 class Footer extends React.Component {
 
   render() {
     const menu1 = this.props.links ? <Menu items={this.props.links.footer1} /> : null;
     const menu2 = this.props.links ? <Menu items={this.props.links.footer2} /> : null;
+    const customClasses = this.props.customClasses;
 
     return (
-      <Wrapper className="container-fluid">
+      <div className={`${customClasses} dc-footer`}>
         <div className="page-footer">
           
-            <Branding>
+            <div className="branding">
               <h2>Open Source Open Data</h2>
               <p>
                 We can only realize the full power of open data when the tools used for its collection, publishing and analysis are also open and transparent.
@@ -30,14 +29,13 @@ class Footer extends React.Component {
                 <a href="https://dkan.slack.com/"><i className="fa fa-slack" aria-hidden="true" /></a>
                 <a href="https://github.com/getdkan"><i className="fa fa-github" aria-hidden="true" /></a>
               </div>
-            </Branding>
-
+            </div>
 
             {menu1}
 
             {menu2}
         </div>
-      </Wrapper>
+      </div>
     );
   }
 }
