+++ src/theme/styles/_variables.scss
@@ -1,3 +1,4 @@
+// Colors.
 $backgroundColor:   #f9fafb;
 $borderColor:       #dce4ef;
 $primaryDark:       #162e51; // blueDark.
@@ -22,3 +23,26 @@ $navBarLinkHoverBack:   #002245;
 $navBarLinkActiveBack:  #305e92;
 $navBarLinkActiveHoverBack: #f2d13c;
 $navBarLinkActiveHover: #FFFFFF;
+$heroBackgroundColor: #0A77BD;
+$heroTextColor:     #FFFFFF;
+$heroHeadingColor:  #FFFFFF;
+$footerBackgroundColor: #162e51;
+$footerText:        #FFFFFF;
+$footerLink:        #FFFFFF;
+$footerLinkHover:   #FFFFFF;
+// Fonts.
+$fontText: 'Rubik','Open Sans','Helvetica Neue',Helvetica,Arial,sans-serif;
+$fontHeading:'Rubik','Open Sans','Helvetica Neue',Helvetica,Arial,sans-serif;
+// Alerts/Messaging.
+$info:          #16A2B7;
+$warning:       #FEC108;
+$danger:        #DB3545;
+$success:       #28A645;
+// File Formats.
+$csvIcon:       #a64300;
+$rdfIcon:       #0b4498;
+$jsonIcon:      #005051;
+$pdfIcon:       #e0051e;
+$xmlIcon:       #5e962c;
+$zipIcon:       #3B0053;
+$dataIcon:      #666666;
